tools/wycv-master/requirements.txt
tqdm
pypinyin
openpyxl
progress
pandas
pycocotools
opencv-python
labelme
numpy
scikit-learn
matplotlib==3.2
pillow
setuptools
xmltodict
pyyaml
imgviz